examples/example-ai-instructor/pyproject.toml
[project]
name = "example-ai-instructor"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "instructor>=1.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "opentelemetry-instrumentation-openai-v2>=2.0b0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-otlp-proto-http>=1.30.0",
]
